+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# [1.3.1] - 2020-10-31
+## Added
+- add event preview image
+- add navigation to event page
+
 # [1.3.0] - 2020-10-31
 ## Added
 - add navigation feature to map